Refactor Dockerfile, docker-compose.yml, and update dependencies
