+++ Dockerfile
@@ -1,7 +1,7 @@
 # Usar una imagen oficial de Python como base
 FROM python:3.9-slim
 
-# Establecer el directorio de trabajo en el contenedor
+
 WORKDIR /app
 
 # Copiar los archivos del proyecto al contenedor
@@ -14,5 +14,8 @@ RUN pip install -r requirements.txt
 # Exponer el puerto 8080
 EXPOSE 8080
 
+
 # Comando para ejecutar la aplicación usando Gunicorn en el puerto 8080
 CMD [ "gunicorn", "--bind", "0.0.0.0:8080", "app.app:app"]
+
+
