+++ requirements.txt
@@ -11,4 +11,4 @@ celery
 redis
 moviepy
 gunicorn
-
+flower
